+++ databases_preview.go
@@ -36,7 +36,7 @@ type DatabaseNG struct {
 	App        App       `json:"app"`
 }
 
-type databaseAPIResponse struct {
+type databaseListItem struct {
 	Database DatabaseNG `json:"database"`
 }
 
@@ -68,25 +68,25 @@ func (c *PreviewClient) DatabaseCreate(ctx context.Context, params DatabaseCreat
 }
 
 func (c *PreviewClient) DatabasesList(ctx context.Context) ([]DatabaseNG, error) {
-	var listRes []databaseAPIResponse
+	var listRes []databaseListItem
 
 	err := c.parent.ScalingoAPI().ResourceList(ctx, databasesResource, nil, &listRes)
 	if err != nil {
 		return nil, errors.Wrap(ctx, err, "list databases")
 	}
 
-	res := make([]DatabaseNG, len(listRes))
+	databasesNG := make([]DatabaseNG, len(listRes))
 
 	for i, apiResponse := range listRes {
-		databaseNG, err := c.populateAPIResponse(ctx, apiResponse)
+		dbNG, err := c.populateAPIResponse(ctx, apiResponse)
 		if err != nil {
-			return nil, errors.Wrap(ctx, err, "populate databaseNG")
+			return nil, errors.Wrap(ctx, err, "populate database NG")
 		}
 
-		res[i] = databaseNG
+		databasesNG[i] = dbNG
 	}
 
-	return res, nil
+	return databasesNG, nil
 }
 
 // DatabaseShow currently uses appID as the database identifier.
@@ -133,35 +133,35 @@ func (c *PreviewClient) searchDatabase(ctx context.Context, appID string) (Datab
 }
 
 // populateAPIResponse populates a DatabaseNG without using the App and Addon from the databases endpoints.
-func (c *PreviewClient) populateAPIResponse(ctx context.Context, apiResponse databaseAPIResponse) (DatabaseNG, error) {
-	var res DatabaseNG
-
-	res.ID = apiResponse.Database.ID
-	res.Name = apiResponse.Database.Name
-	res.ProjectID = apiResponse.Database.ProjectID
-	res.Technology = apiResponse.Database.Technology
-	res.Plan = apiResponse.Database.Plan
+func (c *PreviewClient) populateAPIResponse(ctx context.Context, apiResponse databaseListItem) (DatabaseNG, error) {
+	databaseNG := DatabaseNG{
+		ID:         apiResponse.Database.ID,
+		Name:       apiResponse.Database.Name,
+		ProjectID:  apiResponse.Database.ProjectID,
+		Technology: apiResponse.Database.Technology,
+		Plan:       apiResponse.Database.Plan,
+	}
 
 	addons, err := c.parent.AddonsList(ctx, apiResponse.Database.ID)
 	if err != nil {
-		return res, errors.Wrap(ctx, err, "list addons")
+		return databaseNG, errors.Wrap(ctx, err, "list addons")
 	}
 
 	if len(addons) == 0 {
-		return res, errors.New(ctx, "no addons found for database")
+		return databaseNG, errors.New(ctx, "no addons found for database")
 	}
 
 	app, err := c.parent.AppsShow(ctx, apiResponse.Database.Name)
 	if err != nil {
-		return res, errors.Wrap(ctx, err, "show app")
+		return databaseNG, errors.Wrap(ctx, err, "show app")
 	}
-	res.App = *app
+	databaseNG.App = *app
 
 	database, err := c.parent.DatabaseShow(ctx, apiResponse.Database.ID, addons[0].ID)
 	if err != nil {
-		debug.Printf("Addons probably deleted for app: %+v\n", res.Name)
+		debug.Printf("Addon has been removed from app: %+v\n", databaseNG.Name)
 	}
-	res.Database = &database
+	databaseNG.Database = &database
 
-	return res, nil
+	return databaseNG, nil
 }