+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+* feat(databases): update databases preview service to use database object from API
+
 ## 8.7.1
 
 * feat(events) add database backup events